# perf: Binary search in token store `utils.search`