--- a/lib/source-code/token-store/utils.js
+++ b/lib/source-code/token-store/utils.js
@@ -31,7 +31,7 @@ function getStartLocation(token) {
  */
 exports.search = function search(tokens, location) {
     for (let minIndex = 0, maxIndex = tokens.length - 1; minIndex <= maxIndex;) {
-        const index = (minIndex + maxIndex) / 2 | 0;
+        const index = Math.trunc((minIndex + maxIndex) / 2);
         const token = tokens[index];
 
         if (location <= getStartLocation(token)) {