--- a/lib/source-code/token-store/utils.js
+++ b/lib/source-code/token-store/utils.js
@@ -5,20 +5,6 @@
 "use strict";
 
 //------------------------------------------------------------------------------
-// Helpers
-//------------------------------------------------------------------------------
-
-/**
- * Gets `token.range[0]` from the given token.
- * @param {Node|Token|Comment} token The token to get.
- * @returns {number} The start location.
- * @private
- */
-function getStartLocation(token) {
-    return token.range[0];
-}
-
-//------------------------------------------------------------------------------
 // Exports
 //------------------------------------------------------------------------------
 
@@ -31,10 +17,18 @@ function getStartLocation(token) {
  */
 exports.search = function search(tokens, location) {
     for (let minIndex = 0, maxIndex = tokens.length - 1; minIndex <= maxIndex;) {
-        const index = Math.trunc((minIndex + maxIndex) / 2);
-        const token = tokens[index];
 
-        if (location <= getStartLocation(token)) {
+        /*
+         * Calculate the index in the middle between minIndex and maxIndex.
+         * `| 0` is used to round a fractional value down to the nearest integer: this is similar to
+         * using `Math.trunc()` or `Math.floor()`, but performance tests have shown this method to
+         * be faster.
+         */
+        const index = (minIndex + maxIndex) / 2 | 0;
+        const token = tokens[index];
+        const tokenStartLocation = token.range[0];
+
+        if (location <= tokenStartLocation) {
             if (index === minIndex) {
                 return index;
             }